+++ temporian/core/typing.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import Dict, List, Set, Type, TypeVar, Union
+from typing import Dict, List, Set, Tuple, Type, TypeVar, Union
 from temporian.core.data.dtype import DType
 
 from temporian.core.data.node import EventSetNode
@@ -60,9 +60,44 @@ NodeToEventSetMapping = Union[
 If a dictionary, the mapping is defined by it.
 
 If a single EventSet or a list of EventSets, each EventSet is mapped to their
-own node using [`EventSet.node()`][temporian.EventSet.node], i.e., `[event_set]`
-is equivalent to `{event_set.node() : event_set}`.
+own node using [`EventSet.node()`][temporian.EventSet.node], i.e., `[evset]`
+is equivalent to `{evset.node() : evset}`.
 """
 
-
 TypeOrDType = Union[DType, Type[float], Type[int], Type[str], Type[bool]]
+
+
+IndexKeyItem = Union[int, str, bytes]
+"""One of the values inside an [IndexKey][temporian.core.typing.IndexKey]."""
+
+
+IndexKey = Union[Tuple[IndexKeyItem, ...], IndexKeyItem]
+"""An index key is a tuple of values that identifies a single time sequence
+inside an [`EventSet`][temporian.EventSet].
+
+If, for example, your EventSet is indexed by `"name"` and `"number"`, with the
+values on those being [`"Mark", "Sarah"]` and `[1, 2, 3]` respectively, the
+possible index keys would be `("Mark", 1)`, `("Mark", 2)`, `("Mark", 3)`,
+`("Sarah", 1)`, `("Sarah", 2)`, and `("Sarah", 3)`.
+
+An index key can take the form of a single value (e.g. `"Mark"`) if it is being
+used with an EventSet with a single index. In this case, using `"Mark"` is
+equivalent to using `("Mark",)`.
+"""
+
+IndexKeyList = Union[IndexKey, List[IndexKey]]
+"""A list of [`IndexKeys`][temporian.core.typing.IndexKey].
+
+Auxiliary type to allow receiving a single IndexKey or a list of IndexKeys.
+
+If receiving a single IndexKey, it is equivalent to receiving a list with a
+single IndexKey.
+"""
+
+# Internal
+
+# Internally we only allow normalized index keys, i.e., tuples of bytes and ints
+# IndexKeyItem and IndexKey are user-facing and are normalized before being
+# passed to implementations/serialized/etc.
+NormalizedIndexKeyItem = Union[int, bytes]
+NormalizedIndexKey = Tuple[NormalizedIndexKeyItem, ...]